+++ CHANGELOG.md
@@ -7,6 +7,9 @@ this project uses date-based 'snapshot' version identifiers.
 
 ## [Unreleased]
 
+### Changed
+- Exclude `sourcefiles` entries from file clean-up
+
 ## [2020-03-16]
 
 ### Changed